+++ tests/TestCase/Model/Table/PostsTableTest.php
@@ -136,6 +136,7 @@ class PostsTableTest extends PostsAndPagesTablesTestCase
         $this->assertArrayKeysEqual(['Categories', 'Tags', 'Users'], $query->getContain());
 
         $this->skipIf(!$this->isMySql());
+        $this->skipIfCakeIsLessThan('4.3');
         $this->assertStringEndsWith('FROM `posts` `Posts` INNER JOIN `posts_categories` `Categories` ON `Categories`.`id` = `Posts`.`category_id` INNER JOIN `users` `Users` ON `Users`.`id` = `Posts`.`user_id` ORDER BY `Posts`.`created` DESC', $query->sql());
     }
 
@@ -193,6 +194,7 @@ class PostsTableTest extends PostsAndPagesTablesTestCase
     public function testQueryFromFilter(): void
     {
         $this->skipIf(!$this->isMySql());
+        $this->skipIfCakeIsLessThan('4.3');
         $query = $this->Table->queryFromFilter($this->Table->find(), ['tag' => 'test']);
         $this->assertStringEndsWith('FROM `posts` `Posts` INNER JOIN `posts_tags` `PostsTags` ON `Posts`.`id` = `PostsTags`.`post_id` INNER JOIN `tags` `Tags` ON (`tag` = :c0 AND `Tags`.`id` = `PostsTags`.`tag_id`)', $query->sql());
         $this->assertEquals('test', $query->getValueBinder()->bindings()[':c0']['value']);
@@ -206,17 +208,31 @@ class PostsTableTest extends PostsAndPagesTablesTestCase
     {
         $this->skipIf(!$this->isMySql());
 
-        $query = $this->Table->queryForRelated(4, true);
-        $this->assertStringEndsWith('FROM `posts` `Posts` INNER JOIN `posts_tags` `PostsTags` ON `Posts`.`id` = `PostsTags`.`post_id` INNER JOIN `tags` `Tags` ON (`Tags`.`id` = :c0 AND `Tags`.`id` = `PostsTags`.`tag_id`) WHERE (`Posts`.`active` = :c1 AND `Posts`.`created` <= :c2 AND (`Posts`.`preview`) IS NOT NULL AND `Posts`.`preview` != :c3)', $query->sql());
+        $query = $this->Table->queryForRelated(4);
+        $sql = $query->sql();
         $this->assertEquals(4, $query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertEquals(true, $query->getValueBinder()->bindings()[':c1']['value']);
         $this->assertInstanceof(FrozenTime::class, $query->getValueBinder()->bindings()[':c2']['value']);
         $this->assertEquals([], $query->getValueBinder()->bindings()[':c3']['value']);
 
+        $this->skipIfCakeIsLessThan('4.3');
+        $this->assertStringEndsWith('FROM `posts` `Posts` INNER JOIN `posts_tags` `PostsTags` ON `Posts`.`id` = `PostsTags`.`post_id` INNER JOIN `tags` `Tags` ON (`Tags`.`id` = :c0 AND `Tags`.`id` = `PostsTags`.`tag_id`) WHERE (`Posts`.`active` = :c1 AND `Posts`.`created` <= :c2 AND (`Posts`.`preview`) IS NOT NULL AND `Posts`.`preview` != :c3)', $sql);
+
+    }
+
+    /**
+     * Test for `queryForRelated()` method, without images
+     * @test
+     */
+    public function testQueryForRelatedWithoutImages(): void
+    {
         $query = $this->Table->queryForRelated(4, false);
-        $this->assertStringEndsWith('FROM `posts` `Posts` INNER JOIN `posts_tags` `PostsTags` ON `Posts`.`id` = `PostsTags`.`post_id` INNER JOIN `tags` `Tags` ON (`Tags`.`id` = :c0 AND `Tags`.`id` = `PostsTags`.`tag_id`) WHERE (`Posts`.`active` = :c1 AND `Posts`.`created` <= :c2)', $query->sql());
+        $sql = $query->sql();
         $this->assertEquals(4, $query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertEquals(true, $query->getValueBinder()->bindings()[':c1']['value']);
         $this->assertInstanceof(FrozenTime::class, $query->getValueBinder()->bindings()[':c2']['value']);
+
+        $this->skipIfCakeIsLessThan('4.3');
+        $this->assertStringEndsWith('FROM `posts` `Posts` INNER JOIN `posts_tags` `PostsTags` ON `Posts`.`id` = `PostsTags`.`post_id` INNER JOIN `tags` `Tags` ON (`Tags`.`id` = :c0 AND `Tags`.`id` = `PostsTags`.`tag_id`) WHERE (`Posts`.`active` = :c1 AND `Posts`.`created` <= :c2)', $sql);
     }
 }