+++ tests/TestCase/Model/Table/UsersTableTest.php
@@ -148,9 +148,6 @@ class UsersTableTest extends TableTestCase
         $this->assertTrue($query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertFalse($query->getValueBinder()->bindings()[':c1']['value']);
 
-        $query = $this->Table->find('auth');
-        $this->assertStringEndsWith('FROM `users` `Users` INNER JOIN `users_groups` `Groups` ON `Groups`.`id` = `Users`.`group_id`', $query->sql());
-
         $query = $this->Table->find('banned');
         $this->assertStringEndsWith('FROM `users` `Users` WHERE `banned` = :c0', $query->sql());
         $this->assertTrue($query->getValueBinder()->bindings()[':c0']['value']);
@@ -159,6 +156,10 @@ class UsersTableTest extends TableTestCase
         $this->assertStringEndsWith('FROM `users` `Users` WHERE (`active` = :c0 AND `banned` = :c1)', $query->sql());
         $this->assertFalse($query->getValueBinder()->bindings()[':c0']['value']);
         $this->assertFalse($query->getValueBinder()->bindings()[':c1']['value']);
+
+        $this->skipIfCakeIsLessThan('4.3');
+        $query = $this->Table->find('auth');
+        $this->assertStringEndsWith('FROM `users` `Users` INNER JOIN `users_groups` `Groups` ON `Groups`.`id` = `Users`.`group_id`', $query->sql());
     }
 
     /**