+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ## 2.30 branch
 ### 2.30.4-RC1
 * fixed a large number of deprecations introduced by CakePHP 4.3;
-* added `TestCase::skipIfCakeIsLessThan()` method.
+* added `TestCase::skipIfCakeIsLessThan()` method;
+* it successfully skips tests that generate a different sql syntax.
 
 ### 2.30.3-beta
 * removed the files that interacted with the `cakephp-database-backup` plugin.